remove bugs in view_fines function

+++ Assignment1/src/Include/date_functions.h
@@ -1,4 +1,4 @@
 #pragma once
 int valid_date_checker(int issue_month, int issue_year, int issue_date );
 int count_leap_years(int day, int month, int year);
-int days_calculator(int day_1, int month_1, int year_1, int day_2, int month_2, int year_2);
+long long int days_calculator(int day_1, int month_1, int year_1, int day_2, int month_2, int year_2);
